Crate color

Crate color is an ANSI color package to output colorized or SGR defined output to the standard output. The API can be used in several way, pick one that suits you.

This is a line-by-line port of github.com/fatih/color (v1.19.0). The Go API is preserved verbatim, adjusted only for Rust naming conventions (New → Color::new, SprintFunc → Color::sprint_func, FgRed → FG_RED, …).

Use simple and default helper functions with predefined foreground colors:

use color::{cyan, blue, red, yellow, magenta, hi_green, hi_black, hi_white, vals, NO_ARGS};

cyan("Prints text in cyan.", NO_ARGS);

// a newline will be appended automatically
blue("Prints %s in blue.", &vals!["text"]);

// More default foreground colors..
red("We have red", NO_ARGS);
yellow("Yellow color too!", NO_ARGS);
magenta("And many others ..", NO_ARGS);

// Hi-intensity colors
hi_green("Bright green color.", NO_ARGS);
hi_black("Bright black means gray..", NO_ARGS);
hi_white("Shiny white color!", NO_ARGS);

However, there are times when custom color mixes are required. Below are some examples to create custom color objects and use the print functions of each separate color object.

use color::{vals, Color, BOLD, BG_WHITE, FG_BLUE, FG_CYAN, FG_RED, UNDERLINE};

// Create a new color object
let mut c = Color::new(&[FG_CYAN]);
c.add(&[UNDERLINE]);
c.println(&vals!["Prints cyan text with an underline."]).unwrap();

// Or just add them to New()
let d = Color::new(&[FG_CYAN, BOLD]);
d.printf("This prints bold cyan %s\n", &vals!["too!."]).unwrap();

// Mix up foreground and background colors, create new mixes!
let mut red = Color::new(&[FG_RED]);

let bold_red = red.add(&[BOLD]).clone();
bold_red.println(&vals!["This will print text in bold red."]).unwrap();

let white_background = red.add(&[BG_WHITE]).clone();
white_background.println(&vals!["Red text with White background."]).unwrap();

// Use your own io::Write output
Color::new(&[FG_BLUE]).fprintln(&mut my_writer, &vals!["blue color!"]).unwrap();

let blue = Color::new(&[FG_BLUE]);
blue.fprint(&mut my_writer, &vals!["This will print text in blue."]).unwrap();

You can create print_xxx functions to simplify even more:

use color::{vals, Color, BOLD, FG_GREEN, FG_RED, NO_ARGS};

// Create a custom print function for convenient
let red = Color::new(&[FG_RED]).printf_func();
red("warning", NO_ARGS);
red("error: %s", &vals!["file not found"]);

// Mix up multiple attributes
let notice = Color::new(&[BOLD, FG_GREEN]).println_func();
notice(&vals!["don't forget this..."]);

You can also fprint_xxx functions to pass your own writer:

use color::{vals, Color, BOLD, FG_BLUE, FG_GREEN};

let blue = Color::new(&[FG_BLUE]).fprintf_func();
blue(&mut my_writer, "important notice: %s", &vals!["***"]);

// Mix up with multiple attributes
let success = Color::new(&[BOLD, FG_GREEN]).fprintln_func();
success(&mut my_writer, &vals!["don't forget this..."]);

Or create sprint_xxx functions to mix strings with other non-colorized strings:

use color::{gofmt, vals, Color, BG_GREEN, FG_RED, FG_WHITE, FG_YELLOW};

let yellow = Color::new(&[FG_YELLOW]).sprint_func();
let red = Color::new(&[FG_RED]).sprint_func();

print!("{}", gofmt::sprintf(
    "this is a %s and this is %s.\n",
    &vals![yellow(&vals!["warning"]), red(&vals!["error"])],
));

let info = Color::new(&[FG_WHITE, BG_GREEN]).sprint_func();
print!("{}", gofmt::sprintf("this %s rocks!\n", &vals![info(&vals!["package"])]));

Windows support is enabled by default. All print functions work as intended. However, only for the sprint_xxx functions, users should write to output:

use color::{gofmt, output, vals, green_string, Color, BG_GREEN, FG_WHITE, NO_ARGS};
use std::io::Write;

let mut out = output();
write!(out, "{}", gofmt::sprintf("Windows support: %s", &vals![green_string("PASS", NO_ARGS)])).unwrap();

let info = Color::new(&[FG_WHITE, BG_GREEN]).sprint_func();
write!(out, "{}", gofmt::sprintf("this %s rocks!\n", &vals![info(&vals!["package"])])).unwrap();

Using with existing code is possible. Just use the set function to set the standard output to the given parameters. That way a rewrite of an existing code is not required.

use color::{set, unset, BOLD, FG_MAGENTA, FG_YELLOW};

// Use handy standard colors.
set(&[FG_YELLOW]);

println!("Existing text will be now in Yellow");
println!("This one {}", "too");

unset(); // don't forget to unset

// You can mix up parameters
set(&[FG_MAGENTA, BOLD]);

println!("All text will be now bold magenta.");
unset();

There might be a case where you want to disable color output (for example to pipe the standard output of your app to somewhere else). Color has support to disable colors both globally and for single color definition. For example suppose you have a CLI app and a --no-color bool flag. You can easily disable the color output with:

use color::set_no_color;

let flag_no_color = true;
if flag_no_color {
    set_no_color(true); // disables colorized output
}

You can also disable the color by setting the NO_COLOR environment variable to any value.

It also has support for single color definitions (local). You can disable/enable color output on the fly:

use color::{vals, Color, FG_CYAN};

let mut c = Color::new(&[FG_CYAN]);
c.println(&vals!["Prints cyan text"]).unwrap();

c.disable_color();
c.println(&vals!["This is printed without any color"]).unwrap();

c.enable_color();
c.println(&vals!["This prints again cyan..."]).unwrap();

Modules

gofmt

A port of the subset of Go’s fmt package used by this crate. A port of the subset of Go’s fmt package that fatih/color relies on.

Macros

vals

Builds a Vec<Value> — the Rust equivalent of a Go ...interface{} argument list.

Structs

Attribute

Attribute defines a single SGR Code.

Color

Color defines a custom color object which is defined by SGR parameters.

SharedBuffer

A cheaply clonable, shared, growable buffer.

WriteError

The error half of Go’s (n int, err error) print contract.

Enums

Value

A dynamically typed value, the Rust stand-in for Go’s interface{}.

Writer

The io.Writer interface.

Constants

BG_BLACK

Black background.

BG_BLUE

Blue background.

BG_CYAN

Cyan background.

BG_GREEN

Green background.

BG_HI_BLACK

Hi-intensity black (gray) background.

BG_HI_BLUE

Hi-intensity blue background.

BG_HI_CYAN

Hi-intensity cyan background.

BG_HI_GREEN

Hi-intensity green background.

BG_HI_MAGENTA

Hi-intensity magenta background.

BG_HI_RED

Hi-intensity red background.

BG_HI_WHITE

Hi-intensity white background.

BG_HI_YELLOW

Hi-intensity yellow background.

BG_MAGENTA

Magenta background.

BG_RED

Red background.

BG_WHITE

White background.

BG_YELLOW

Yellow background.

BLINK_RAPID

Rapid blink.

BLINK_SLOW

Slow blink.

BOLD

Bold / increased intensity.

CONCEALED

Concealed / hidden.

CROSSED_OUT

Crossed out / strike-through.

FAINT

Faint / decreased intensity.

FG_BLACK

Black foreground.

FG_BLUE

Blue foreground.

FG_CYAN

Cyan foreground.

FG_GREEN

Green foreground.

FG_HI_BLACK

Hi-intensity black (gray) foreground.

FG_HI_BLUE

Hi-intensity blue foreground.

FG_HI_CYAN

Hi-intensity cyan foreground.

FG_HI_GREEN

Hi-intensity green foreground.

FG_HI_MAGENTA

Hi-intensity magenta foreground.

FG_HI_RED

Hi-intensity red foreground.

FG_HI_WHITE

Hi-intensity white foreground.

FG_HI_YELLOW

Hi-intensity yellow foreground.

FG_MAGENTA

Magenta foreground.

FG_RED

Red foreground.

FG_WHITE

White foreground.

FG_YELLOW

Yellow foreground.

ITALIC

Italic.

NO_ARGS

The empty argument list, a convenience for the very common f(format) / f(format, ...) split in the helper functions.

RESET

Resets all attributes.

RESET_BLINKING

Turns off blinking.

RESET_BOLD

Turns off bold/faint.

RESET_CONCEALED

Turns off conceal.

RESET_CROSSED_OUT

Turns off crossed out.

RESET_ITALIC

Turns off italic.

RESET_REVERSED

Turns off reverse video.

RESET_UNDERLINE

Turns off underline.

REVERSE_VIDEO

Reverse video (swap foreground/background).

UNDERLINE

Underline.

Functions

bg_rgb

Port of BgRGB(r, g, b int) *Color.

black

Black is a convenient helper function to print with black foreground. A newline is appended to format by default.

black_string

BlackString is a convenient helper function to return a string with black foreground.

blue

Blue is a convenient helper function to print with blue foreground. A newline is appended to format by default.

blue_string

BlueString is a convenient helper function to return a string with blue foreground.

cyan

Cyan is a convenient helper function to print with cyan foreground. A newline is appended to format by default.

cyan_string

CyanString is a convenient helper function to return a string with cyan foreground.

equals

Port of (*Color).Equals including its nil receiver handling.

error

Reads the global Error writer.

green

Green is a convenient helper function to print with green foreground. A newline is appended to format by default.

green_string

GreenString is a convenient helper function to return a string with green foreground.

hi_black

HiBlack is a convenient helper function to print with hi-intensity black foreground. A newline is appended to format by default.

hi_black_string

HiBlackString is a convenient helper function to return a string with hi-intensity black foreground.

hi_blue

HiBlue is a convenient helper function to print with hi-intensity blue foreground. A newline is appended to format by default.

hi_blue_string

HiBlueString is a convenient helper function to return a string with hi-intensity blue foreground.

hi_cyan

HiCyan is a convenient helper function to print with hi-intensity cyan foreground. A newline is appended to format by default.

hi_cyan_string

HiCyanString is a convenient helper function to return a string with hi-intensity cyan foreground.

hi_green

HiGreen is a convenient helper function to print with hi-intensity green foreground. A newline is appended to format by default.

hi_green_string

HiGreenString is a convenient helper function to return a string with hi-intensity green foreground.

hi_magenta

HiMagenta is a convenient helper function to print with hi-intensity magenta foreground. A newline is appended to format by default.

hi_magenta_string

HiMagentaString is a convenient helper function to return a string with hi-intensity magenta foreground.

hi_red

HiRed is a convenient helper function to print with hi-intensity red foreground. A newline is appended to format by default.

hi_red_string

HiRedString is a convenient helper function to return a string with hi-intensity red foreground.

hi_white

HiWhite is a convenient helper function to print with hi-intensity white foreground. A newline is appended to format by default.

hi_white_string

HiWhiteString is a convenient helper function to return a string with hi-intensity white foreground.

hi_yellow

HiYellow is a convenient helper function to print with hi-intensity yellow foreground. A newline is appended to format by default.

hi_yellow_string

HiYellowString is a convenient helper function to return a string with hi-intensity yellow foreground.

magenta

Magenta is a convenient helper function to print with magenta foreground. A newline is appended to format by default.

magenta_string

MagentaString is a convenient helper function to return a string with magenta foreground.

new

Port of New(value ...Attribute) *Color.

new_colorable_stderr

Port of colorable.NewColorableStderr().

new_colorable_stdout

Port of colorable.NewColorableStdout().

no_color

Reads the global NoColor flag.

output

Reads the global Output writer.

red

Red is a convenient helper function to print with red foreground. A newline is appended to format by default.

red_string

RedString is a convenient helper function to return a string with red foreground.

rgb

Port of RGB(r, g, b int) *Color.

set

Port of Set(p ...Attribute) *Color.

set_error

Assigns the global Error writer — the equivalent of color.Error = w.

set_no_color

Assigns the global NoColor flag — the equivalent of color.NoColor = v.

set_output

Assigns the global Output writer — the equivalent of color.Output = w.

set_stderr_available

Sets whether standard error is considered available.

set_stdout_available

Sets whether standard output is considered available.

std_err

Port of stdErr.

std_out

Port of stdOut.

stderr_available

Reports whether standard error is considered available (os.Stderr != nil).

stdout_available

Reports whether standard output is considered available (os.Stdout != nil).

stdout_is_terminal

Port of stdoutIsTerminal.

unset

Port of Unset().

white

White is a convenient helper function to print with white foreground. A newline is appended to format by default.

white_string

WhiteString is a convenient helper function to return a string with white foreground.

yellow

Yellow is a convenient helper function to print with yellow foreground. A newline is appended to format by default.

yellow_string

YellowString is a convenient helper function to return a string with yellow foreground.

Type Aliases

Result

Result of the printing functions: the byte count on success, or the byte count plus the error on failure.