color::cache_key

Struct CacheKey

#[repr(transparent)]
pub struct CacheKey<T>(pub T);

A key usable in a hashmap to compare the bit representation types containing colors.

See the module level docs for more information.

Tuple Fields
0: T
Implementations

impl<T> CacheKey<T>

pub fn new(value: T) -> Self

Create a new CacheKey.

All fields are public, so the struct constructor can also be used.

pub fn into_inner(self) -> T

Get the inner value.