color::cache_key

Trait BitEq

pub trait BitEq {
    // Required method
    fn bit_eq(&self, other: &Self) -> bool;
}

An equivalence relation for types which normally wouldn’t have one, implemented using a bitwise comparison for floating point values.

See the docs on Eq for more information.

Useful for creating caches based on exact values. See the module level docs for more information.

Required Methods

fn bit_eq(&self, other: &Self) -> bool

Returns true if self is equal to other.

This need not use the semantically natural comparison operation for the type; indeed floating point types should implement this by comparing bit values.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.