color::color

Function fast_round_to_u8

fn fast_round_to_u8(a: f32) -> u8

Fast rounding of f32 to integer u8, rounding ties up.

Targeting x86, f32::round calls out to libc roundf. Even if that call were inlined, it is branchy, which would make it relatively slow. The following is faster, and on the range 0-255 almost correct*. AArch64 has dedicated rounding instructions so does not need this optimization, but the following is still fast.

The only input where the output differs from a.round() as u8 is 0.49999997.