color

Module colorspace

Structs

A98Rgb

🌌 The Adobe RGB (1998) color space.

Aces2065_1

🌌 The ACES2065-1 color space.

AcesCg

🌌 The ACEScg color space.

DisplayP3

🌌 The Display P3 color space, often used for wide-gamut displays.

Hsl

🌌 The HSL color space

Hwb

🌌 The HWB color space

Lab

🌌 The CIELAB color space

Lch

🌌 The cylindrical version of the Lab color space.

LinearSrgb

🌌 The linear-light RGB color space with sRGB primaries.

Oklab

🌌 The Oklab color space, intended to be a perceptually uniform color space.

Oklch

🌌 The cylindrical version of the Oklab color space.

ProphotoRgb

🌌 The ProPhoto RGB color space.

Rec2020

🌌 The Rec. 2020 color space.

Srgb

🌌 The standard RGB color space.

XyzD50

🌌 The CIE XYZ color space with a 2° observer and a reference white of D50.

XyzD65

🌌 The CIE XYZ color space with a 2° observer and a reference white of D65.

Enums

ColorSpaceLayout

The layout of a color space, particularly the hue component.

Constants

EPSILON 🔒

KAPPA 🔒

LAB_SRGB_TO_XYZ 🔒

LAB_XYZ_TO_SRGB 🔒

OKLAB_LAB_TO_LMS 🔒

OKLAB_LMS_TO_LAB 🔒

OKLAB_LMS_TO_SRGB 🔒

OKLAB_SRGB_TO_LMS 🔒

Traits

ColorSpace

The main trait for color spaces.

Functions

hsl_to_rgb 🔒

Convert HSL to RGB.

hwb_to_rgb 🔒

Convert HWB to RGB.

lab_to_lch 🔒

Rectangular to cylindrical conversion.

lch_to_lab 🔒

Cylindrical to rectangular conversion.

lin_to_srgb 🔒

rgb_to_hsl 🔒

Convert RGB to HSL.

rgb_to_hwb 🔒

Convert RGB to HWB.

srgb_to_lin 🔒