color

Module parse

Parse CSS4 color

Structs

Parser 🔒

Enums

Mode 🔒

Whether or not we are parsing modern or legacy mode syntax.

ParseError

Error type for parse errors.

Value 🔒

A parsed value.

Constants

LOWERCASE_BUF_SIZE 🔒

Functions

color_from_4bit_hex 🔒

color_from_components 🔒

get_4bit_hex_channels 🔒

Parse 4-bit color channels from a hex-encoded string.

hex_from_ascii_byte 🔒

make_lowercase 🔒

If the string contains any uppercase characters, make a lowercase copy in the provided buffer space.

parse_color

Parse a color string in CSS syntax into a color.

parse_color_prefix

Parse a color string prefix in CSS syntax into a color.