color::parse

Function parse_color

pub fn parse_color(s: &str) -> Result<DynamicColor, ParseError>

Parse a color string in CSS syntax into a color.

This parses the entire string; trailing characters cause an ExpectedEndOfString parse error. Leading and trailing whitespace are ignored. See also parse_color_prefix.

Errors

Tries to return a suitable error for any invalid string, but may be a little lax on some details.