color

Struct AlphaColor

#[repr(transparent)]
pub struct AlphaColor<CS> {
    pub components: [f32; 4],
    pub cs: PhantomData<CS>,
}

A color with an alpha channel.

A color in a color space known at compile time, with an alpha channel.

The color channels are straight, i.e., they are not premultiplied by the alpha channel. See PremulColor for a color type with color channels premultiplied by the alpha channel.

See OpaqueColor for a discussion of arithmetic traits and interpolation.

Fields
components: [f32; 4]

The components, which may be manipulated directly.

The interpretation of the first three components depends on the color space. The fourth component is separate alpha.

cs: PhantomData<CS>

The color space.

Implementations

impl<CS: ColorSpace> AlphaColor<CS>

pub const BLACK: Self

A black color.

More comprehensive pre-defined colors are available in the color::palette module.

pub const TRANSPARENT: Self

A transparent color.

This is a black color with full alpha.

More comprehensive pre-defined colors are available in the color::palette module.

pub const WHITE: Self

A white color.

This value is specific to the color space.

More comprehensive pre-defined colors are available in the color::palette module.

pub const fn new(components: [f32; 4]) -> Self

Create a new color from the given components.

pub const fn split(self) -> (OpaqueColor<CS>, f32)

Split into opaque and alpha components.

This function is the inverse of OpaqueColor::with_alpha.

pub const fn with_alpha(self, alpha: f32) -> Self

Set the alpha channel.

This replaces the existing alpha channel. To scale or or otherwise modify the existing alpha channel, use AlphaColor::multiply_alpha or AlphaColor::map.

let c = color::palette::css::GOLDENROD.with_alpha(0.5);
assert_eq!(0.5, c.split().1);

pub const fn discard_alpha(self) -> OpaqueColor<CS>

Split out the opaque components, discarding the alpha.

This is a shorthand for calling split.

pub fn convert<TargetCs: ColorSpace>(self) -> AlphaColor<TargetCs>

Convert a color into a different color space.

pub const fn premultiply(self) -> PremulColor<CS>

Convert a color to the corresponding premultiplied form.

pub(crate) fn difference(self, other: Self) -> f32

Difference between two colors by Euclidean metric.

pub fn lerp_rect(self, other: Self, t: f32) -> Self

Linearly interpolate colors, without hue fixup.

This method produces meaningful results in rectangular color spaces, or if hue fixup has been applied.

pub fn lerp(self, other: Self, t: f32, direction: HueDirection) -> Self

Linearly interpolate colors, with hue fixup if needed.

pub const fn multiply_alpha(self, rhs: f32) -> Self

Multiply alpha by the given factor.

pub fn scale_chroma(self, scale: f32) -> Self

Scale the chroma by the given amount.

See ColorSpace::scale_chroma for more details.

pub fn map(self, f: impl Fn(f32, f32, f32, f32) -> [f32; 4]) -> Self

Map components.

pub fn map_in<TargetCS: ColorSpace>( self, f: impl Fn(f32, f32, f32, f32) -> [f32; 4], ) -> Self

Map components in a given color space.

pub fn map_lightness(self, f: impl Fn(f32) -> f32) -> Self

Map the lightness of the color.

In a color space that naturally has a lightness component, map that value. Otherwise, do the mapping in Oklab. The lightness range is normalized so that 1.0 is white. That is the normal range for Oklab but differs from the range in Lab, Lch, and Hsl.

Examples

use color::{AlphaColor, Lab};

let color = AlphaColor::<Lab>::new([40., 4., -17., 1.]);
let lighter = color.map_lightness(|l| l + 0.2);
let expected = AlphaColor::<Lab>::new([60., 4., -17., 1.]);

assert!(lighter.premultiply().difference(expected.premultiply()) < 1e-4);

pub fn map_hue(self, f: impl Fn(f32) -> f32) -> Self

Map the hue of the color.

In a color space that naturally has a hue component, map that value. Otherwise, do the mapping in Oklch. The hue is in degrees.

Examples

use color::{AlphaColor, Oklab};

let color = AlphaColor::<Oklab>::new([0.5, 0.2, -0.1, 1.]);
let complementary = color.map_hue(|h| (h + 180.) % 360.);
let expected = AlphaColor::<Oklab>::new([0.5, -0.2, 0.1, 1.]);

assert!(complementary.premultiply().difference(expected.premultiply()) < 1e-4);

pub fn to_rgba8(self) -> Rgba8

Convert the color to sRGB if not already in sRGB, and pack into 8 bit per component integer encoding.

The RGBA components are mapped from the floating point range of 0.0-1.0 to the integer range of 0-255. Component values outside of this range are saturated to 0 or 255.

Implementation note

This performs almost-correct rounding to be fast on both x86 and AArch64 hardware. Within the saturated output range of this method, 0-255, there is a single color component value where results differ: 0.0019607842. This method maps that component to integer value 1; it would more precisely be mapped to 0.

impl AlphaColor<Srgb>

pub const fn from_rgba8(r: u8, g: u8, b: u8, a: u8) -> Self

Create a color from 8-bit rgba values.

Note: for conversion from the Rgba8 type, just use the From trait.

pub const fn from_rgb8(r: u8, g: u8, b: u8) -> Self

Create a color from 8-bit rgb values with an opaque alpha.

Note: for conversion from the Rgba8 type, just use the From trait.