color

Struct Color

pub struct Color { /* private fields */ }

Color defines a custom color object which is defined by SGR parameters.

Port of Go’s

type Color struct {
    params  []Attribute
    noColor *bool
}

*bool becomes Option<bool>: None defers to the global no_color flag, Some(v) is an explicit per-color override installed by Color::disable_color / Color::enable_color.

Implementations

impl Color

pub fn new(value: &[Attribute]) -> Color

Port of New(value ...Attribute) *Color.

New returns a newly created color object.

pub fn params(&self) -> &[Attribute]

The SGR parameters that make up this color.

pub fn add_rgb(&mut self, r: i32, g: i32, b: i32) -> &mut Self

Port of (*Color).AddRGB.

AddRGB is used to chain foreground RGB SGR parameters. Use as many as parameters to combine and create custom color objects. Example: .Add(34, 0, 12).Add(255, 128, 0).

pub fn add_bg_rgb(&mut self, r: i32, g: i32, b: i32) -> &mut Self

Port of (*Color).AddBgRGB.

AddBgRGB is used to chain background RGB SGR parameters. Use as many as parameters to combine and create custom color objects. Example: .Add(34, 0, 12).Add(255, 128, 0).

pub fn set(&self) -> &Self

Port of (*Color).Set.

Set sets the SGR sequence.

pub fn set_writer<W: Write + ?Sized>(&self, w: &mut W) -> &Self

Port of (*Color).SetWriter.

SetWriter is used to set the SGR sequence with the given io.Writer. This is a low-level function, and users should use the higher-level functions, such as Fprint, Print, etc.

pub fn unset_writer<W: Write + ?Sized>(&self, w: &mut W)

Port of (*Color).UnsetWriter.

UnsetWriter resets all escape attributes and clears the output with the given io.Writer. Usually should be called after SetWriter.

pub fn add(&mut self, value: &[Attribute]) -> &mut Self

Port of (*Color).Add.

Add is used to chain SGR parameters. Use as many as parameters to combine and create custom color objects. Example: Add(color.FgRed, color.Underline).

Like the Go method this mutates the receiver in place and returns it for chaining.

pub fn fprint<W: Write + ?Sized>(&self, w: &mut W, a: &[Value]) -> Result<usize>

Port of (*Color).Fprint.

Fprint formats using the default formats for its operands and writes to w. Spaces are added between operands when neither is a string. It returns the number of bytes written and any write error encountered.

pub fn print(&self, a: &[Value]) -> Result<usize>

Port of (*Color).Print.

Print formats using the default formats for its operands and writes to standard output. Spaces are added between operands when neither is a string. It returns the number of bytes written and any write error encountered. This is the standard fmt.Print() method wrapped with the given color.

pub fn fprintf<W: Write + ?Sized>( &self, w: &mut W, format: &str, a: &[Value], ) -> Result<usize>

Port of (*Color).Fprintf.

Fprintf formats according to a format specifier and writes to w. It returns the number of bytes written and any write error encountered.

pub fn printf(&self, format: &str, a: &[Value]) -> Result<usize>

Port of (*Color).Printf.

Printf formats according to a format specifier and writes to standard output. It returns the number of bytes written and any write error encountered. This is the standard fmt.Printf() method wrapped with the given color.

pub fn fprintln<W: Write + ?Sized>( &self, w: &mut W, a: &[Value], ) -> Result<usize>

Port of (*Color).Fprintln.

Fprintln formats using the default formats for its operands and writes to w. Spaces are always added between operands and a newline is appended.

pub fn println(&self, a: &[Value]) -> Result<usize>

Port of (*Color).Println.

Println formats using the default formats for its operands and writes to standard output. Spaces are always added between operands and a newline is appended. It returns the number of bytes written and any write error encountered. This is the standard fmt.Print() method wrapped with the given color.

pub fn sprint(&self, a: &[Value]) -> String

Port of (*Color).Sprint.

Sprint is just like Print, but returns a string instead of printing it.

pub fn sprintln(&self, a: &[Value]) -> String

Port of (*Color).Sprintln.

Sprintln is just like Println, but returns a string instead of printing it.

pub fn sprintf(&self, format: &str, a: &[Value]) -> String

Port of (*Color).Sprintf.

Sprintf is just like Printf, but returns a string instead of printing it.

pub fn fprint_func(&self) -> impl Fn(&mut dyn Write, &[Value])

Port of (*Color).FprintFunc.

FprintFunc returns a new function that prints the passed arguments as colorized with Fprint().

pub fn print_func(&self) -> impl Fn(&[Value])

Port of (*Color).PrintFunc.

PrintFunc returns a new function that prints the passed arguments as colorized with Print().

pub fn fprintf_func(&self) -> impl Fn(&mut dyn Write, &str, &[Value])

Port of (*Color).FprintfFunc.

FprintfFunc returns a new function that prints the passed arguments as colorized with Fprintf().

pub fn printf_func(&self) -> impl Fn(&str, &[Value])

Port of (*Color).PrintfFunc.

PrintfFunc returns a new function that prints the passed arguments as colorized with Printf().

pub fn fprintln_func(&self) -> impl Fn(&mut dyn Write, &[Value])

Port of (*Color).FprintlnFunc.

FprintlnFunc returns a new function that prints the passed arguments as colorized with Fprintln().

pub fn println_func(&self) -> impl Fn(&[Value])

Port of (*Color).PrintlnFunc.

PrintlnFunc returns a new function that prints the passed arguments as colorized with Println().

pub fn sprint_func(&self) -> impl Fn(&[Value]) -> String

Port of (*Color).SprintFunc.

SprintFunc returns a new function that returns colorized strings for the given arguments with fmt.Sprint(). Useful to put into or mix into other string.

put := New(FgYellow).SprintFunc()
fmt.Fprintf(color.Output, "This is a %s", put("warning"))

pub fn sprintf_func(&self) -> impl Fn(&str, &[Value]) -> String

Port of (*Color).SprintfFunc.

SprintfFunc returns a new function that returns colorized strings for the given arguments with fmt.Sprintf().

pub fn sprintln_func(&self) -> impl Fn(&[Value]) -> String

Port of (*Color).SprintlnFunc.

SprintlnFunc returns a new function that returns colorized strings for the given arguments with fmt.Sprintln().

pub fn disable_color(&mut self)

Port of (*Color).DisableColor.

DisableColor disables the color output. Useful to not change any existing code and still being able to output. Can be used for flags like --no-color. To enable back use EnableColor() method.

pub fn enable_color(&mut self)

Port of (*Color).EnableColor.

EnableColor enables the color output. Use it in conjunction with DisableColor(). Otherwise, this method has no side effects.

pub fn equals(&self, c2: &Color) -> bool

Port of (*Color).Equals.

Equals returns a boolean value indicating whether two colors are equal.

Attribute multisets are compared: order does not matter but duplicate counts do.