Crate color

A simple representation of the standard RGB color model.

Structs

Color
This structure represents a color value in the standard RGB color model, comprised of red, blue, green, and transparency components.

Constants

BLACK

BLUE

BROWN

CYAN

DARK_GRAY

GRAY

GREEN

LIGHT_BLUE

LIGHT_CYAN

LIGHT_GRAY

LIGHT_GREEN

MAGENTA

PINK

RED

TRANSPARENT

WHITE

YELLOW