color::color

Function fixup_hue

fn fixup_hue(h1: f32, h2: &mut f32, direction: HueDirection)

Fixup hue based on specified hue direction.

Reference: 12.4 of CSS Color 4 spec

Note that this technique has been tweaked to only modify the second hue. The rationale for this is to support multiple gradient stops, for example in a spline. Apply the fixup to successive adjacent pairs.

In addition, hues outside [0, 360) are supported, with a resulting hue difference always in [-360, 360].